refactor: change operand/option 'optional' logic to 'required' logic
This is just the opposite logic for both. I felt the concept of `optional`
didn't really look good for `Option` and if I changed it for `Option` it
would need to be changed for `Operand` as well, so I did. I feel that
`required` makes more sense at a glance, anyway.